requirements.txt
alembic>=1.16.0
fastapi>=0.115.0
openai>=1.0.0
pydantic-settings>=2.8.0
sqlalchemy>=2.0.39
uvicorn>=0.34.0
httpx>=0.28.0
pytest>=8.3.0
psycopg[binary]>=3.2.0

